refactor: replace DM Sans font import with @fontsource and update Storybook configuration
- Updated .storybook/preview.tsx to use @fontsource/dm-sans for font loading.
- Added storybook.css to define CSS variables for font compatibility.
- Removed unused DM_Sans import and related code.
- Updated package.json to include @fontsource/dm-sans dependency.
- Updated pnpm-lock.yaml to reflect new dependency.
- Refactored Content and MediaBlock stories to use @storybook/jest and @storybook/testing-library for testing.
