+++ package.json
@@ -80,6 +80,7 @@
     "@chromatic-com/storybook": "^4.1.3",
     "@eslint/css": "^0.14.1",
     "@eslint/eslintrc": "^3.3.3",
+    "@fontsource/dm-sans": "^5.2.8",
     "@next/eslint-plugin-next": "16.1.1",
     "@poupe/eslint-plugin-tailwindcss": "^0.2.4",
     "@radix-ui/react-slider": "^1.3.6",
@@ -88,9 +89,9 @@
     "@storybook/addon-onboarding": "^10.1.11",
     "@storybook/addon-vitest": "^10.1.11",
     "@storybook/jest": "^0.2.3",
+    "@storybook/nextjs-vite": "^10.1.11",
     "@storybook/react": "^10.1.11",
     "@storybook/react-vite": "^10.1.11",
-    "@storybook/test": "^8.6.15",
     "@storybook/testing-library": "^0.2.2",
     "@tailwindcss/postcss": "^4.1.18",
     "@tailwindcss/typography": "^0.5.19",
